refactor: изменить адресацию контроллеров.
make MarkerImpl абстрактным

delete MarkerImpl
