+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -3,7 +3,7 @@ package ru.yandex.practicum.filmorate.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import ru.yandex.practicum.filmorate.annotation.ReleaseDate;
-import ru.yandex.practicum.filmorate.sirvice.Marker;
+import ru.yandex.practicum.filmorate.validation.Marker;
 import javax.validation.constraints.*;
 import java.time.LocalDate;
