+++ src/main/resources/application.yml
@@ -1,7 +1,6 @@
----
 server:
   port: 8080
----
+
 logging:
   level:
     org:
